Create Changing and giving permission in a department
In a fictional organization, I work with their research team. Part of my task here is to ensure users on this team are authorized with the appropriate permissions as this helps keep the system secure. 
I examined existing permissions on the file system. I determined if the permissions match the authorization that should be given. When they did not match, i modified the permissions to authorize the appropriate users and remove any unauthorized access.


I made use of the Ls -l command to view permissions given to each owner type as shown in the command below (i.e. User, a group, and other employees). There are hidden files in this company too and i used the Ls -la to check who has permission to view those files. .Project_x.txt was the hidden file I was able to expose. Ls -la is used to view both hidden and visible permissions while Ls -a is used for only hidden files only.  I did an act called authorization as I did not just allow everyone access to the files. E.g for Project_k.txt,  i made sure that other employees do not have permission to write to the file by using the command chmod o-w project_k.txt. The chmod command changed the permission with the argument o-w (O standing for others and W standing for write and the -(hyphen) meaning that others should be deprived of writing to the file. This creates order and balances in your workplace.